Uploaded v1 of GA2.zip
I'm assuming we're intended to have the three separate class files and then there will be a main. i haven't made edits to them yet, save for copying out the candy class and putting it in it's own .cpp file. The main file currently has the test bench they gave us in the assignment.